+++ database/seeds/DatabaseSeeder.php
@@ -12,8 +12,6 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        $this->call(BorrowingStatusesSeeder::class);
-        $this->command->info('Borrowing statuses table seeded.');
         $this->call(InventoryItemStatusesSeeder::class);
         $this->command->info('Inventory item statuses table seeded.');
         $this->call(UserRolesSeeder::class);