+++ routes/web.php
@@ -8,6 +8,8 @@ Route::get('/', function () {
 Route::get('/new-borrowing', 'NewBorrowingController@index');
 Route::post('/new-borrowing', 'NewBorrowingController@store');
 Route::get('/end-borrowing', 'EndBorrowingController@index');
+Route::post('/end-borrowing/returned', 'EndBorrowingController@updateReturned');
+Route::post('/end-borrowing/lost', 'EndBorrowingController@updateLost');
 Route::get('/borrowings-history', 'BorrowingsHistoryController@index')->name('borrowings-history');
 Route::get('/view-inventory', 'ViewInventoryController@index');
 Route::get('/edit-inventory', 'EditInventoryController@index');