+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.6.0] - Unreleased
+## [0.6.0] - 2019-09-19
 
 ### Added
 
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Selection modal
 * End borrowing controller
     * Borrowings request.
+    * Borrowing + Inventory item update.
 
 ### Changed
 
@@ -36,6 +37,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Seeds
     * User role for user seeding wasn't a constant from ```UserRole```.
+    
+### Removed
+
+* Database / Model
+    * Removed borrowing statuses in favor of a boolean ```finished```.
 
 ## [0.5.0] - 2018-09-16